tests: move fixtures to a subfolder
