+++ test/fixtures/amd.js
