+++ test/fixtures/cjsExportLazy.js
