+++ test/fixtures/cjsMixedImport.js
