+++ test/fixtures/commonjs-requiremain.js
