+++ test/fixtures/coreModules.js
