+++ test/fixtures/es6.js
