+++ test/fixtures/es6DynamicImport.js
