+++ test/fixtures/es6MixedExportLazy.js
