+++ test/fixtures/es6MixedImport.js
