+++ test/fixtures/es6NoImport.js
