+++ test/fixtures/es6WithError.js
