+++ test/fixtures/es7.js
