+++ test/fixtures/internalNodePrefix.js
