+++ test/fixtures/typescript.ts
